generate new ui
