--- gui/README.md
@@ -1,29 +1,0 @@
-# Scte35Js
-
-This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 7.3.6.
-
-## Deploying changes to gh-pages
-
-Run:
-
-`ng build --prod --base-href /scte35-js/`
-
-This should generate fresh copies of the files in `/docs`. Push these changes up.
-
-Check out your newly deployed changes at `https://<user-name>.github.io/scte35-js/` (and once the changes are pushed upstream at `https://comcast.github.io/scte35-js/`)
-
-## Development server
-
-Run `ng serve` for a dev server. Navigate to `http://localhost:4200/`. The app will automatically reload if you change any of the source files.
-
-## Code scaffolding
-
-Run `ng generate component component-name` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module`.
-
-## Build
-
-Run `ng build` to build the project. The build artifacts will be stored in the `dist/` directory. Use the `--prod` flag for a production build.
-
-## Further help
-
-To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI README](https://github.com/angular/angular-cli/blob/master/README.md).
