--- gui/src/app/app.component.html
@@ -1,65 +1,0 @@
-<!-- Toolbar -->
-<div class="toolbar" role="banner">
-  <img width="40" alt="Logo" src="http://cdn.comcast.com/learn/-/media/common/favicon/favicon.ico" />
-
-  <span>SCTE35-JS PARSER DEMO</span>
-
-</div>
-
-<div class="content" role="main">
-
-  <div class="scte35-settings">
-    <label id="radio-group-label">
-      <h3>SCTE-35 Format Type</h3>
-    </label>
-    <mat-radio-group aria-labelledby="radio-group-label" class="radio-group" [(ngModel)]="scte35format">
-      <mat-radio-button class="radio-button" *ngFor="let format of formats" [value]="format">
-        {{format}}
-      </mat-radio-button>
-    </mat-radio-group>
-
-
-    <form class="form">
-      <mat-form-field class="full-width">
-        <mat-label>SCTE35 TAG</mat-label>
-        <input matInput [formControl]="scte35FormControl" [errorStateMatcher]="matcher"
-          placeholder="Ex. /DBGAAET8J+pAP/wBQb+AAAAAAAwAi5DVUVJQAErgX+/CR9TSUdOQUw6OGlTdzllUWlGVndBQUFBQUFBQUJCQT09NwMDaJ6RZQ==">
-        <mat-error *ngIf="scte35FormControl.hasError('required')">
-          A valid SCTE35 is <strong>required</strong>
-        </mat-error>
-      </mat-form-field>
-    </form>
-
-    <button mat-raised-button color="accent" (click)="parse()">PARSE</button>
-
-    <mat-card *ngIf="displayObject !== undefined" class="scte35-card">
-      <mat-card-header>
-        <div mat-card-avatar class="example-header-image"></div>
-        <mat-card-title>Parsed SCTE-35</mat-card-title>
-      </mat-card-header>
-      <mat-card-content>
-        <ngx-json-viewer *ngIf="!raw" [json]="displayObject"></ngx-json-viewer>
-        <div class="raw-viewer" *ngIf="raw">{{rawObject}}</div>
-
-      </mat-card-content>
-      <mat-card-actions>
-        <button *ngIf="!raw" mat-raised-button color="primary" (click)="viewRaw()">VIEW RAW</button>
-
-        <button *ngIf="raw" mat-raised-button color="primary" (click)="viewPretty()">VIEW PRETTY</button>
-
-      </mat-card-actions>
-
-    </mat-card>
-
-
-
-
-    <!-- Footer -->
-    <svg id="clouds" alt="Gray Clouds Background" xmlns="http://www.w3.org/2000/svg" width="2611.084" height="485.677"
-      viewBox="0 0 2611.084 485.677">
-      <path id="Path_39" data-name="Path 39"
-        d="M2379.709,863.793c10-93-77-171-168-149-52-114-225-105-264,15-75,3-140,59-152,133-30,2.83-66.725,9.829-93.5,26.25-26.771-16.421-63.5-23.42-93.5-26.25-12-74-77-130-152-133-39-120-212-129-264-15-54.084-13.075-106.753,9.173-138.488,48.9-31.734-39.726-84.4-61.974-138.487-48.9-52-114-225-105-264,15a162.027,162.027,0,0,0-103.147,43.044c-30.633-45.365-87.1-72.091-145.206-58.044-52-114-225-105-264,15-75,3-140,59-152,133-53,5-127,23-130,83-2,42,35,72,70,86,49,20,106,18,157,5a165.625,165.625,0,0,0,120,0c47,94,178,113,251,33,61.112,8.015,113.854-5.72,150.492-29.764a165.62,165.62,0,0,0,110.861-3.236c47,94,178,113,251,33,31.385,4.116,60.563,2.495,86.487-3.311,25.924,5.806,55.1,7.427,86.488,3.311,73,80,204,61,251-33a165.625,165.625,0,0,0,120,0c51,13,108,15,157-5a147.188,147.188,0,0,0,33.5-18.694,147.217,147.217,0,0,0,33.5,18.694c49,20,106,18,157,5a165.625,165.625,0,0,0,120,0c47,94,178,113,251,33C2446.709,1093.793,2554.709,922.793,2379.709,863.793Z"
-        transform="translate(142.69 -634.312)" fill="#eee" />
-    </svg>
-
-  </div>
