--- gui/src/app/app.module.ts
@@ -1,30 +1,0 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { AppComponent } from './app.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatRadioModule } from '@angular/material/radio';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatInputModule } from '@angular/material/input';
-import { MatButtonModule } from '@angular/material/button';
-import { NgxJsonViewerModule } from 'ngx-json-viewer';
-import { MatCardModule } from '@angular/material/card';
-
-@NgModule({
-  declarations: [
-    AppComponent
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    MatRadioModule,
-    FormsModule,
-    MatInputModule,
-    ReactiveFormsModule,
-    MatButtonModule,
-    NgxJsonViewerModule,
-    MatCardModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
-export class AppModule { }
