--- ui/src/app/demo-page/demo-page.component.html
@@ -1,25 +1,0 @@
-<mat-card>
-  <mat-card-title>SCTE35-JS Demo</mat-card-title>
-  <mat-card-content fxLayout="column" fxLayoutAlign="center stretch">
-    <div fxLayout="column" fxLayoutAlign="center stretch">
-      <mat-form-field>
-        <input matInput placeholder="Payload" [(ngModel)]="payload">
-      </mat-form-field>
-    </div>
-    <div fxLayout="row" fxLayoutAlign="space-around center">
-      <button mat-button mat-raised-button color="primary" (click)="parsePayload()">Parse</button>
-      <button mat-button mat-raised-button color="primary" (click)="showRawJson = true" *ngIf="parsedObject !== undefined && !showRawJson">
-        View Raw JSON
-      </button>
-      <button mat-button mat-raised-button color="primary" (click)="showRawJson = false" *ngIf="parsedObject !== undefined && showRawJson">
-        Hide Raw JSON
-      </button>
-    </div>
-    <div fxLayout="column" fxLayoutAlign="center center" *ngIf="parsedObject !== undefined">
-      <ngx-json-viewer [json]="parsedObject"></ngx-json-viewer>
-    </div>
-    <mat-card *ngIf="showRawJson">
-      <pre style="word-break: break-all; white-space: normal;">{{parsedObjectString}}</pre>
-    </mat-card>
-  </mat-card-content>
-</mat-card>
