--- ui/src/app/material/material.module.ts
@@ -1,83 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import {
-  MatButtonModule,
-  MatCheckboxModule,
-  MatToolbarModule,
-  MatExpansionModule,
-  MatCardModule,
-  MatFormFieldModule,
-  MatInputModule,
-  MatStepperModule,
-  MatSnackBarModule,
-  MatListModule,
-  MatSlideToggleModule,
-  MatIconModule,
-  MatTooltipModule,
-  MatMenuModule,
-  MatSelectModule,
-  MatSidenavModule,
-  MatDialogModule,
-  MatTabsModule,
-  MatGridListModule,
-  MatTableModule,
-  MatDividerModule,
-  MatAutocompleteModule,
-  MatProgressSpinnerModule
-} from '@angular/material';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    MatToolbarModule,
-    MatExpansionModule,
-    MatCardModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatStepperModule,
-    MatSnackBarModule,
-    MatListModule,
-    MatSlideToggleModule,
-    MatIconModule,
-    MatTooltipModule,
-    MatMenuModule,
-    MatSelectModule,
-    MatSidenavModule,
-    MatDialogModule,
-    MatTabsModule,
-    MatGridListModule,
-    MatTableModule,
-    MatDividerModule,
-    MatAutocompleteModule,
-    MatProgressSpinnerModule
-  ],
-  exports: [
-    MatButtonModule,
-    MatCheckboxModule,
-    MatToolbarModule,
-    MatExpansionModule,
-    MatCardModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatStepperModule,
-    MatSnackBarModule,
-    MatListModule,
-    MatSlideToggleModule,
-    MatIconModule,
-    MatTooltipModule,
-    MatMenuModule,
-    MatSelectModule,
-    MatSidenavModule,
-    MatDialogModule,
-    MatTabsModule,
-    MatGridListModule,
-    MatTableModule,
-    MatDividerModule,
-    MatAutocompleteModule,
-    MatProgressSpinnerModule
-  ],
-  declarations: []
-})
-export class MaterialModule { }
